+++ bot/src/main/java/edu/java/bot/config/ClientsConfig.java
@@ -1,22 +1,19 @@
 package edu.java.bot.config;
 
-import edu.java.bot.httpClients.ScrapperHttpClient;
-import org.springframework.beans.factory.annotation.Autowired;
+import edu.java.bot.webClients.ScrapperClient;
+import edu.java.bot.webClients.ScrapperWebClient;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
+@RequiredArgsConstructor
 public class ClientsConfig {
     private final ApplicationConfig applicationConfig;
 
-    @Autowired
-    public ClientsConfig(ApplicationConfig applicationConfig) {
-        this.applicationConfig = applicationConfig;
-    }
-
     @Bean
-    public ScrapperHttpClient githubClient() {
-        return new ScrapperHttpClient(applicationConfig.scrapperUrl());
+    public ScrapperClient scrapperClient() {
+        return new ScrapperWebClient(applicationConfig.scrapperUrl());
     }
 
 }
