--- bot/src/main/java/edu/java/bot/dto/request/LinkRequest.java
@@ -1,9 +1,0 @@
-package edu.java.bot.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-import java.net.URI;
-
-public record LinkRequest(
-    @NotNull
-    URI link) {
-}
