+++ bot/src/main/java/edu/java/bot/dto/request/client/LinkRequest.java
@@ -1,0 +1,9 @@
+package edu.java.bot.dto.request.client;
+
+import jakarta.validation.constraints.NotNull;
+import java.net.URI;
+
+public record LinkRequest(
+    @NotNull
+    URI link) {
+}
