--- bot/src/main/java/edu/java/bot/dto/response/LinkResponse.java
@@ -1,8 +1,0 @@
-package edu.java.bot.dto.response;
-
-import java.net.URI;
-
-public record LinkResponse(Long id,
-                           URI url) {
-
-}
