--- bot/src/main/java/edu/java/bot/dto/response/ListLinksResponse.java
@@ -1,5 +1,0 @@
-package edu.java.bot.dto.response;
-
-public record ListLinksResponse(LinkResponse[] links,
-                                Integer size) {
-}
