+++ bot/src/main/java/edu/java/bot/dto/response/client/LinkResponse.java
@@ -1,0 +1,8 @@
+package edu.java.bot.dto.response.client;
+
+import java.net.URI;
+
+public record LinkResponse(Long id,
+                           URI url) {
+
+}
