+++ bot/src/main/java/edu/java/bot/dto/response/client/ListLinksResponse.java
@@ -1,0 +1,5 @@
+package edu.java.bot.dto.response.client;
+
+public record ListLinksResponse(LinkResponse[] links,
+                                Integer size) {
+}
