+++ bot/src/main/java/edu/java/bot/exceptions/ExceptionApiHandler.java
@@ -1,0 +1,44 @@
+package edu.java.bot.exceptions;
+
+import edu.java.bot.dto.response.ApiErrorResponse;
+import java.util.Arrays;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class ExceptionApiHandler {
+
+    @ExceptionHandler(ResourceNotFound.class)
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    public ApiErrorResponse chatNotFoundException(ResourceNotFound ex) {
+
+        return new ApiErrorResponse(
+            "Ресурс не существует",
+            HttpStatus.NOT_FOUND.toString(),
+            "Resource doesn't exist",
+            ex.getMessage(),
+            Arrays.stream(ex.getStackTrace())
+                .map(StackTraceElement::toString)
+                .toArray(String[]::new)
+        );
+    }
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ApiErrorResponse handleValidationExceptions(MethodArgumentNotValidException ex) {
+
+        return new ApiErrorResponse(
+            "Некорректные параметры запроса",
+            ex.getStatusCode().toString(),
+            ex.getBody().getTitle(),
+            ex.getMessage(),
+            Arrays.stream(ex.getStackTrace())
+                .map(StackTraceElement::toString)
+                .toArray(String[]::new)
+        );
+
+    }
+}
