+++ bot/src/main/java/edu/java/bot/exceptions/ResourceNotFound.java
@@ -1,0 +1,9 @@
+package edu.java.bot.exceptions;
+
+public class ResourceNotFound extends RuntimeException {
+    public static final String MSG = "Указанный ресурс не был найден";
+
+    public ResourceNotFound() {
+        super(MSG);
+    }
+}
