--- scrapper/src/main/java/edu/java/scrapper/dto/request/LinkRequest.java
@@ -1,9 +1,0 @@
-package edu.java.scrapper.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-import java.net.URI;
-
-public record LinkRequest(
-    @NotNull
-    URI link) {
-}
