+++ scrapper/src/main/java/edu/java/scrapper/dto/request/client/LinkUpdateRequest.java
@@ -1,0 +1,16 @@
+package edu.java.scrapper.dto.request.client;
+
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.PositiveOrZero;
+import java.net.URI;
+
+public record LinkUpdateRequest(@NotNull
+                                @PositiveOrZero
+                                Long id,
+                                @NotNull
+                                URI url,
+                                String description,
+                                @NotNull
+                                Long[] tgChatIds) {
+
+}
