+++ scrapper/src/main/java/edu/java/scrapper/dto/request/controller/LinkRequest.java
@@ -1,0 +1,9 @@
+package edu.java.scrapper.dto.request.controller;
+
+import jakarta.validation.constraints.NotNull;
+import java.net.URI;
+
+public record LinkRequest(
+    @NotNull
+    URI link) {
+}
