--- scrapper/src/main/java/edu/java/scrapper/dto/response/GitUserResponse.java
@@ -1,8 +1,0 @@
-package edu.java.scrapper.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import java.time.OffsetDateTime;
-
-public record GitUserResponse(@JsonProperty("full_name") String name,
-                              @JsonProperty("updated_at") OffsetDateTime updatedAt) {
-}
