--- scrapper/src/main/java/edu/java/scrapper/dto/response/LinkResponse.java
@@ -1,7 +1,0 @@
-package edu.java.scrapper.dto.response;
-
-import java.net.URI;
-
-public record LinkResponse(Long id, URI url) {
-
-}
