--- scrapper/src/main/java/edu/java/scrapper/dto/response/ListLinksResponse.java
@@ -1,4 +1,0 @@
-package edu.java.scrapper.dto.response;
-
-public record ListLinksResponse(LinkResponse[] links, Integer size) {
-}
