+++ scrapper/src/main/java/edu/java/scrapper/dto/response/controller/LinkResponse.java
@@ -1,0 +1,7 @@
+package edu.java.scrapper.dto.response.controller;
+
+import java.net.URI;
+
+public record LinkResponse(Long id, URI url) {
+
+}
