+++ scrapper/src/main/java/edu/java/scrapper/dto/response/controller/ListLinksResponse.java
@@ -1,0 +1,4 @@
+package edu.java.scrapper.dto.response.controller;
+
+public record ListLinksResponse(LinkResponse[] links, Integer size) {
+}
