--- scrapper/src/main/java/edu/java/scrapper/exceptions/LinkNotFound.java
@@ -1,9 +1,0 @@
-package edu.java.scrapper.exceptions;
-
-public class LinkNotFound extends RuntimeException {
-    public static final String MSG = "Указанная ссылка не была найдена";
-
-    public LinkNotFound() {
-        super(MSG);
-    }
-}
