+++ scrapper/src/main/java/edu/java/scrapper/exceptions/ResourceNotFound.java
@@ -1,0 +1,9 @@
+package edu.java.scrapper.exceptions;
+
+public class ResourceNotFound extends RuntimeException {
+    public static final String MSG = "Указанный ресурс не был найден";
+
+    public ResourceNotFound() {
+        super(MSG);
+    }
+}
