--- scrapper/src/main/java/edu/java/scrapper/httpClients/BotClient.java
@@ -1,7 +1,0 @@
-package edu.java.scrapper.httpClients;
-
-import edu.java.scrapper.dto.request.LinkUpdateRequest;
-
-public interface BotClient {
-    void sendUpdate(LinkUpdateRequest linkUpdate);
-}
