--- scrapper/src/main/java/edu/java/scrapper/httpClients/GitClient.java
@@ -1,10 +1,0 @@
-package edu.java.scrapper.httpClients;
-
-import edu.java.scrapper.dto.response.GitUserResponse;
-
-public interface GitClient {
-    GitUserResponse fetchUser(String user);
-
-    GitUserResponse fetchUserRepo(String user, String repo);
-
-}
