--- scrapper/src/main/java/edu/java/scrapper/httpClients/StackClient.java
@@ -1,7 +1,0 @@
-package edu.java.scrapper.httpClients;
-
-import edu.java.scrapper.dto.response.StackUserResponse;
-
-public interface StackClient {
-    StackUserResponse fetchQuestion(String id);
-}
