+++ scrapper/src/main/java/edu/java/scrapper/webClients/BotClient.java
@@ -1,0 +1,7 @@
+package edu.java.scrapper.webClients;
+
+import edu.java.scrapper.dto.request.client.LinkUpdateRequest;
+
+public interface BotClient {
+    void sendUpdate(LinkUpdateRequest linkUpdate);
+}
