+++ scrapper/src/main/java/edu/java/scrapper/webClients/BotWebClient.java
@@ -1,0 +1,32 @@
+package edu.java.scrapper.webClients;
+
+import edu.java.scrapper.dto.request.client.LinkUpdateRequest;
+import edu.java.scrapper.dto.response.ApiErrorResponse;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.MediaType;
+import org.springframework.web.reactive.function.BodyInserters;
+import org.springframework.web.reactive.function.client.WebClient;
+import reactor.core.publisher.Mono;
+
+public class BotWebClient implements BotClient {
+
+    private final WebClient webClient;
+
+    public BotWebClient(String url) {
+        this.webClient = WebClient.builder().baseUrl(url).build();
+    }
+
+    @Override
+    public void sendUpdate(LinkUpdateRequest linkUpdate) {
+        webClient.post()
+            .uri("/updates")
+            .contentType(MediaType.APPLICATION_JSON)
+            .body(BodyInserters.fromValue(linkUpdate))
+            .retrieve()
+            .onStatus(HttpStatusCode::is4xxClientError, response -> response.bodyToMono(ApiErrorResponse.class)
+                .flatMap(errorResponse -> Mono.error(new RuntimeException(String.valueOf(errorResponse)))))
+            .bodyToMono(Void.class)
+            .block();
+    }
+
+}
