+++ scrapper/src/main/java/edu/java/scrapper/webClients/GitClient.java
@@ -1,0 +1,10 @@
+package edu.java.scrapper.webClients;
+
+import edu.java.scrapper.dto.response.client.GitUserResponse;
+
+public interface GitClient {
+    GitUserResponse fetchUser(String user);
+
+    GitUserResponse fetchUserRepo(String user, String repo);
+
+}
