+++ scrapper/src/main/java/edu/java/scrapper/webClients/StackClient.java
@@ -1,0 +1,7 @@
+package edu.java.scrapper.webClients;
+
+import edu.java.scrapper.dto.response.client.StackUserResponse;
+
+public interface StackClient {
+    StackUserResponse fetchQuestion(String id);
+}
